refact: Move all functoions imported from developer_tools into appropriate modules
